Refactor code and modify the structure of the project
