+++ src/main/java/org/fundacionjala/core/api/RequestManager.java
@@ -6,17 +6,11 @@
  * accordance with the terms of the license agreement you entered into
  * with Jalasoft.
  */
-package org.fundacionjala.pivotal;
+package org.fundacionjala.core.api;
 
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
-import io.restassured.specification.RequestSpecification;
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-import org.json.simple.parser.ParseException;
-
-import java.io.FileReader;
-import java.io.IOException;
+import org.fundacionjala.pivotal.Authentication;
 import java.util.Map;
 
 
@@ -32,46 +26,12 @@ public final class RequestManager {
     private RequestManager() { }
 
     /**
-     * The method load a config.json file and return Request specification.
-     * @return returns a RequestSpecification object.
-     * @throws IOException throws input /output exception.
-     * @throws ParseException throws parse exception.
-     */
-    public static RequestSpecification getRequestSpecification() {
-        RequestSpecification requestSpecification = null;
-        final String fileName = "./configJson/config.json";
-        JSONObject jsonObject = getJsonObject(fileName);
-
-        //Reading the String
-        String token = (String) jsonObject.get("x-trackerToken");
-        requestSpecification = new io.restassured.builder.RequestSpecBuilder()
-                .setBaseUri("https://www.pivotaltracker.com/services/v5")
-                .addHeader("X-TrackerToken", token)
-                .build();
-
-         requestSpecification
-                .log().all();
-        return requestSpecification;
-    }
-
-    public static JSONObject getJsonObject(final String fileName) {
-        JSONObject jsonObject;
-        JSONParser parser = new JSONParser();
-        try (FileReader reader = new FileReader(fileName)) {
-            jsonObject = (JSONObject) parser.parse(reader);
-        } catch (IOException | ParseException e) {
-            throw new AWT04exception(1, e.getMessage());
-        }
-        return jsonObject;
-    }
-
-    /**
      * Sets get method path.
      * @param path string path to use url.
      * @return response object.
      */
     public static Response get(final String path) {
-        Response response = given(getRequestSpecification())
+        Response response = given(Authentication.getRequestSpecification())
                 .log().all()
                 .when()
                 .get(path);
@@ -87,7 +47,7 @@ public final class RequestManager {
      */
 
     public static Response post(final String path, final String json) {
-        Response response = given(getRequestSpecification())
+        Response response = given(Authentication.getRequestSpecification())
                 .when()
                 .contentType(ContentType.JSON)
                 .body(json)
@@ -97,7 +57,7 @@ public final class RequestManager {
     }
 
     public static Response post(final String path, final Map<String, String> json) {
-        Response response = given(getRequestSpecification())
+        Response response = given(Authentication.getRequestSpecification())
                 .params(json)
                 .when()
                 .post(path);
@@ -113,7 +73,7 @@ public final class RequestManager {
      */
 
     public static Response put(final String path, final String json) {
-        Response response = given(getRequestSpecification())
+        Response response = given(Authentication.getRequestSpecification())
                 .when()
                 .body(json)
                 .put(path);
@@ -127,7 +87,7 @@ public final class RequestManager {
      * @return response object.
      */
     public static Response delete(final String path) {
-        Response response = given(getRequestSpecification())
+        Response response = given(Authentication.getRequestSpecification())
                 .when()
                 .delete(path);
         response.then().log().all();
@@ -139,8 +99,8 @@ public final class RequestManager {
      * @param path string path to use url.
      * @return response object.
      */
-    public static Response setPatch(final String path) {
-        Response response = given(getRequestSpecification())
+    public static Response patch(final String path) {
+        Response response = given(Authentication.getRequestSpecification())
                 .when()
                 .patch(path);
         response.then().log().all();
