+++ src/test/java/org/fundacionjala/pivotal/steps/Runner.java
@@ -3,7 +3,7 @@ package org.fundacionjala.pivotal.steps;
 import io.cucumber.testng.AbstractTestNGCucumberTests;
 import io.cucumber.testng.CucumberOptions;
 import io.restassured.response.Response;
-import org.fundacionjala.pivotal.RequestManager;
+import org.fundacionjala.core.api.RequestManager;
 import org.testng.annotations.BeforeTest;
 
 import java.util.List;
